refactor: put request interceptor in dedicated file
